Introduced `PlotlyJS` #54 Squashed commit of the following:
commit f9e2007b406da3fa8d6915adfaa7c687168619c3
Author: Mizu49 <mizuki.abe.r4@dc.tohoku.ac.jp>
Date:   Sat Oct 23 21:20:54 2021 +0000

    specify argument type for functions in `PlotRecipe` #54

commit 589a0ba3eed5b9a83781deac9cf0c41e27e77c71
Author: Mizu49 <mizuki.abe.r4@dc.tohoku.ac.jp>
Date:   Sat Oct 23 20:58:42 2021 +0000

    separate `PlotRecipe`'s dependent functions #54

commit ec9e2815fd17d82013b497563eaff27ba62179d5
Author: Mizu49 <mizuki.abe.r4@dc.tohoku.ac.jp>
Date:   Sat Oct 23 20:27:57 2021 +0000

    add setting for plot backend in PlotRecipe #54

commit d9342ec9fb69ce2977db08fe93656e17b4218b62
Author: Mizu49 <mizuki.abe.r4@dc.tohoku.ac.jp>
Date:   Sat Oct 23 18:22:07 2021 +0000

    add `plotlyjs()` in module `PlotRecipe` #54

commit 6c34003cfd82b382f2c4406ef5a378c6b1720367
Author: Mizu49 <mizuki.abe.r4@dc.tohoku.ac.jp>
Date:   Sat Oct 23 18:09:13 2021 +0000

    add `plotlyJS` in `Project.toml` #54
